#19 opening/StudyInfo 컴포넌트 분리
